code del & query other file ok #9
